Support multiple interfaces per coclass
This meant implementing proper QueryInterface support, which is able to
give out different vtable pointers depending on the requested IID.

Also because the caller will now use different vtable pointers when
performing the calls, the implementing methods must know how to find the
original CoClass address based on the vtable address coming in as the
self_void parameter.

Furthermore we need to give the different interfaces proper IID GUIDs.
Previously we only had one GUID per coclass, which acted as both IID and
CLSID (which wasn't really proper according to spec). Now we have
various attributes to use for CLSID and IID:

- com_library: This is unchanged. It still lists all the available
CoClasses.
- com_interface: This attribute is attached on either a trait or an impl
and creates a vtable struct for the methods. The attribute also assigns
an IID for this vtable.
- com_visible: This attribute is mostly unchanged in semantics. However
the attribute is now attached on structs and lists all interfaces that
the struct implements. The expansion of the attribute will result in the
primary QueryInterface, AddRef and Release implementations as well as
the outbound CoClass struct.
- com_impl: This attribute will implement the interface-specific
delegate methods. These include the methods defined in the interface
that will delegate to the Rust Fns as well as the custom IUnknown
methods that resolve the CoClass address and then delegate to the
primary IUnknown methods expanded from [com_visible].

Fix README example