Move most of ClassFactory implementation to com_runtime
Attempt to make as much of the code static as possible. While the auto generated code is surprisingly easy to write (quote_* <3), it's still pain to debug.

The more static code we have the easier the debugging. Also thanks to the generic types and awesomeness of Rust, static code doesn't mean dynamic dispatch.