Implement String parameters
Rust's std::string::String can now be used as a parameter and return type. These are represented as BSTR's in the COM api.

Due to the UTF-8 encoding requirements of std::string::String, the use of these parameters involve string allocations.

This commit also breaks the platform independence as we need to allocate the BSTRs using the Windows ::SysAllocString* APIs. The plan for the future is to use SysAllocString in Windows and Rust alloc cross platform. When Rust alloc is used, the memory needs to be released using Rust's memory utils. This will be done by adding a built in IStringAllocator/CoStringAllocator types which in Windows delegate to SysAllocString/etc. and elsewhere use Rust's Box (or similar) utilities. This allows cross platform deallocation.