Allow COM interface pointers as parameters
Implemented ability to pass COM interface pointers as parameters and
invoke the interfaces from Rust. Currently this is limited to traits and
the methods available on the interfaces are the public methods on those
traits.

In the future we will want to support passing in implicit struct impl
interfaces and also the ability to query the underlying Rust struct from
these interfaces so we can better hide the implementation details.

The implementation for this involved a lot of refactoring. As we now
have two places where we need to convert between Rust parameters/return
values and COM parameters/return values, a new concept of
`ComMethodInfo` was introduced.

The `ComMethodInfo` is a higher level data structure that contains the
resolved parameter/return value handling. It is meant to replace the
various `quote::Tokens` based `get_method_args`/`get_method_rvalues`
utility methods. Previously these methods were responsible for taking
the method signature, figuring out what kind of parameters/return values
the COM method needed and then outputting the `Tokens` to be inserted
into the code.

The new `ComMethodInfo` encapsulates the parameter/return value
resolution and leaves it up to the caller to figure out how these
parameters and return values are used in the final code. This did result
in more complex methods in the `intercom-attributes` attribute
expansion, but it felt clearer to have these methods responsible for the
final code structure instead of breaking that code generation in various
different files (and even crates).