Update README.md
Removed the status and expansions. The status issues listed in the README.md
are now in the GitHub issue tracker while the expansions are somehow
documented in the intercom-attributes tests. While the test sources are not
as documentation-oriented as the README.md expansions were, at least they are
kept up to date through testing while the README.md expansions were outdated.