Support tuples inside Result<(..), E>
In a way this commit is one step forward, two steps back. The tuple handling
was implemented on top of the methodinfo/returnhandlers infrastructure so this
felt like a good moment to transfer all intercom-attributes and the parsing
in intercom-utils on top of the methodinfo/returnhandlers implementation.

However this altered intercom-utils in such a way that handling ComItf/ComRc
return values became more difficult. Due to this this commit breaks those
features currently. These would be fixable by throwing more code at the
problem - however pursuing the strong interface concept would make this
easier.

So the utils are left broken for now. Following commits will implement
the strong interfaces in Rust and fix the utils.

Closes #20