Clarify ComItf/ComRc/ComBox/ComStruct
Lots of changes to how these types are handled. Naming for them is not
final now that we introduced ComStruct to replace ComBox in its use.

Anyway! Here goes with the current naming.

## ComBox

ComBox is a memory layout struct. It defines the way our COM classes are
laid in memory. It shouldn't be needed by user APIs.

## ComStruct

Technically ComStruct is a handle to a ComBox in memory. From user
perspective ComStruct is the way to construct COM-enabled structs. It
should replace ComBox in all of its previous use cases (save some bits
in the expanded code).

ComStruct<T> requires T to be a `CoClass`, but NOT a ´ComInterface´.
The previous way of constructing every struct as a `ComRc<T>` required
`T` to be a `ComInterface` as well, which isn't always the case.

## ComItf

Wraps a COM pointer. As an improvement the pointer may point to a COM-
interface defined as a trait or an implicit impl of a CoClass.

Although currently the COM-interface defined as an implicit impl must be
tied to a local CoClass. It is NOT okay to implement such an interface
in the COM server and pass one to an Intercom component. Intercom will
assume the pointer will point to a local ComBox and cast it as such.

The `create` functionality was removed from ComItf as unsafe. It's now
defined for ComRc, which implements safe reference counting.

## ComRc

A reference counted version of a ComItf.

Closes #28, Closes #5