Allow parsing #![com_library] on crate level
This can't actually be used for now. While RFC 2126 improves on the #41430
situation, the underlying issue is that emitting code from the crate-level
attributes causes ICEs and won't export the symbols.

I'm still committing this changes as I ended up writing the code and we do
want the feature enabled in the future once it works (hopefully it does?).