Discovery of libraries implementing COM classes on POSIX
All Rust COM libraries implement a method called
"IntercomListClassObjects". This function returns all the COM classes
that the Rust COM library implements.

On runtime intercom searches
for all the libraries linked to the process that implement this function
and then associates classes returned by the function with the library.

Currently the discovery relies on "glibc". On platforms that do not ship
 glibc the discovery can be implemented with the help of "ldd".