Ensure "test_lib" gets loaded into the test processes
The tests themselves do not directly refer to any functions within
the Rust COM library. Because of this the linker may decide
to remove all the references the library from the final exectuable.
This then prevents the automatic disecovery of the COM clases,

However, I think the workaround here of calling the
test_lib::Descriptor::is_available() method is a bit too surprising
to be the final implementation.


I have two competing proposals for solving this issue. I prefer
the option B at the moment.

a)
Force a call to the Rust COM library in the generated .cpp file
during the static initialization. The .cpp file is generated from the
Rust COM library and the users of the library are supposed
to include and compile this file within their project anyway.

The disadvantage of this approach is that we cannot react to missing
libraries at runtime. Also I think that on Windows it is a bad idea
to call a function of another DLL during static initialization. Static
initializers for a dll are run inside DllMain and the "loader lock"
is held during this time.


b)
Register the name of the library to the generic "intercom" library
in the generated .cpp file during the static initialization. The
generic "intercom" part can then lookup for the library dynamically
when the creation of an unknown class is requested.

The disadvantage of this approach is that the "intercom" cannot be
a header only library. To avoid the "static initialization order
fiasco" we need a unique compilation unit for the library registry.
But even without this requirement we may want to implement the
"intercom" as a separate library as this would reduce compile times.
The library should be static to avoid issues with "loader lock".