Move `Deref for ComItf<T>` into intercom
Previously the `Deref` was implemented through the attributes. Now that we
have it in the intercom crate itself we don't need `fundamental` feature
anymore.

To drop the requirement for `fundamental` some `From`, etc. implementations
were also removed for now. These should be implemented within the intercom
crate in a similar manner to the `Deref` in the future.

They weren't used in the tests either so I was too tired to figure out what
the big benefit with them was. I believe they did some interface querying,
etc. but we can live without that _for now_.