Introduce raw::InterfacePtr<T> as the raw interface pointer
Previously we used ComItf<T> for this, but given that's a fat type now we
need a new type that is COM compatible. Note that InterfacePtr<T> does not
make any guarantees on which type system the T-pointer represents. This
needs to be handled externally such as InterfacePtrs returned from Automation
type system interfaces represent Automation type system interfaces themselves.