Implement From<ComStruct> for ComItf using HasInterface
HasInterface is an empty marker trait that defines which interfaces a
com_class implements. In the future this marker trait will probably get
specific vtbl offset methods for instantly converting a com_class into
a specific interface.

ESPECIALLY as currently the implementation only fills ONE of the ts-ptrs
while it really should fill both of them.