Removed direction member from TypeContext
The direction member in the TypeContext struct was very difficult to work with when I tried to implement structs which are essentially nested types. The type handles of the individual fields of the structs should not be aware of the direction of struct. The direction is now given as a parameter during code generation,for the type handler.