Make cpp-raw tests succeed in Windows
We are still leaking memory like crazy. The biggest unresolved issue is our
handling of CString return values. Currently we are not re-allocating these,
but this might be problematic as the CStrings do not define how they allocate
the memory and require the dealloc to happen by passing the pointer back.

I guess we'll need one of the following:
- Implement IAllocator::FreeCString similar to our BString variant.
- Implement our own CeeString that is guaranteed to be allocated in a way
  compatible with our current IAllocator::Free
- Always clone CStrings on top of *mut c_char that is allocated with
  IAllocator.